Update image handling

+++ lib/image.functions.php
@@ -59,14 +59,45 @@ function RemoveImage($imageId)
 	}
 }
 
+function CanProcessImages()
+{
+	return function_exists('imagejpeg')
+		&& function_exists('imagecreatetruecolor')
+		&& function_exists('imagecopyresampled');
+}
+
+function CanReadImageType($type)
+{
+	switch ((int)$type) {
+		case 1:
+			return function_exists('imagecreatefromgif');
+		case 2:
+			return function_exists('imagecreatefromjpeg');
+		case 3:
+			return function_exists('imagecreatefrompng');
+		default:
+			return false;
+	}
+}
+
 function ConvertToJpeg($file_src, $file_dst)
 {
+	if (!CanProcessImages()) {
+		return false;
+	}
 
 	if (is_file($file_dst)) {
 		unlink($file_dst); //  remove old images if present
 	}
 
-	list($w_src, $h_src, $type) = getimagesize($file_src);
+	$imageInfo = getimagesize($file_src);
+	if ($imageInfo === false) {
+		return false;
+	}
+	list($w_src, $h_src, $type) = $imageInfo;
+	if (!CanReadImageType($type)) {
+		return false;
+	}
 
 	switch ($type) {
 		case 1:   //   gif -> jpg
@@ -82,18 +113,33 @@ function ConvertToJpeg($file_src, $file_dst)
 			break;
 	}
 
+	if (!$img_src) {
+		return false;
+	}
+
 	imagejpeg($img_src, $file_dst);    //  save new image
 	imagedestroy($img_src);
+	return true;
 }
 
 function CreateThumb($file_src, $file_dst, $w_dst, $h_dst)
 {
+	if (!CanProcessImages()) {
+		return false;
+	}
 
 	if (is_file($file_dst)) {
 		unlink($file_dst); //  remove old images if present
 	}
 
-	list($w_src, $h_src, $type) = getimagesize($file_src);
+	$imageInfo = getimagesize($file_src);
+	if ($imageInfo === false) {
+		return false;
+	}
+	list($w_src, $h_src, $type) = $imageInfo;
+	if (!CanReadImageType($type)) {
+		return false;
+	}
 
 	// create new dimensions, keeping aspect ratio
 	$ratio = $w_src / $h_src;
@@ -116,11 +162,20 @@ function CreateThumb($file_src, $file_dst, $w_dst, $h_dst)
 			break;
 	}
 
+	if (!$img_src) {
+		return false;
+	}
+
 	$img_dst = imagecreatetruecolor($w_dst, $h_dst);  //  resample
+	if (!$img_dst) {
+		imagedestroy($img_src);
+		return false;
+	}
 
 	imagecopyresampled($img_dst, $img_src, 0, 0, 0, 0, $w_dst, $h_dst, $w_src, $h_src);
 	imagejpeg($img_dst, $file_dst);    //  save new image
 
 	imagedestroy($img_src);
 	imagedestroy($img_dst);
+	return true;
 }
